Add `use experimental :pack`
SOCKS.pm6 is using experimental pack/unpack rakudo feature which
needs to be enabled with a pragma. Rakudo versions ≤ 2018.05 used to
give a run time error on first use, but starting with 2018.06 it will
be a compile time error. Tests were never hitting the lines
with `pack` which is why they were green previously.